+++ cli/src/connect/wizard/prop_mapping.ts
@@ -2,7 +2,7 @@ import { ReactProjectInfo } from '../project'
 import { ComponentTypeSignature } from '../../react/parser'
 import { FigmaRestApi } from '../figma_rest_api'
 import { PropMapping } from '../parser_executable_types'
-import { Searcher } from 'fast-fuzzy'
+import { MatchData, Searcher } from 'fast-fuzzy'
 import { BaseCommand } from '../../commands/connect'
 import { logger } from '../../common/logging'
 import { Intrinsic, IntrinsicKind, ValueMapping, ValueMappingKind } from '../../connect/intrinsics'
@@ -37,17 +37,15 @@ export function generateValueMapping(
  * actually be related, so we return null if no suitable mapping
  */
 function generateIntrinsic({
-  propSignature: propSignatureWithOptionalModifier,
-  figmaPropName,
+  propSignature,
+  figmaPropName: figmaPropNameWithNodeId,
   figmaPropDef,
 }: {
   propSignature: string
   figmaPropName: string
   figmaPropDef: FigmaRestApi.ComponentPropertyDefinition
 }): Intrinsic | null {
-  const propSignature = propSignatureWithOptionalModifier.startsWith('?')
-    ? propSignatureWithOptionalModifier.substring(1)
-    : propSignatureWithOptionalModifier
+  const figmaPropName = stripNodeIdFromPropertyName(figmaPropNameWithNodeId)
   if (propSignature === 'string' && figmaPropDef.type === FigmaRestApi.ComponentPropertyType.Text) {
     return {
       kind: IntrinsicKind.String,
@@ -74,15 +72,77 @@ function generateIntrinsic({
     propSignature.includes(' | ') &&
     figmaPropDef.type === FigmaRestApi.ComponentPropertyType.Variant
   ) {
-    return {
-      kind: IntrinsicKind.Enum,
-      args: {
-        figmaPropName,
-        valueMapping: generateValueMapping(propSignature, figmaPropDef),
-      },
+    const valueMapping = generateValueMapping(propSignature, figmaPropDef)
+    // Only valuable if some values were mapped
+    if (Object.keys(valueMapping).length > 0) {
+      return {
+        kind: IntrinsicKind.Enum,
+        args: {
+          figmaPropName,
+          valueMapping,
+        },
+      }
     }
   }
   return null
+}
+
+function stripNodeIdFromPropertyName(propertyName: string) {
+  return propertyName.replace(/#\d+:\d+/, '')
+}
+
+const DELIMITERS_REGEX = /[\s-_]/g
+
+export enum MatchableNameTypes {
+  Property,
+  VariantValue,
+  // ChildLayer, // TODO
+}
+
+type MatchableName = {
+  type: MatchableNameTypes
+  name: string
+  variantProperty?: string // Only for VariantValue
+}
+
+/**
+ * Builds a map of all properties and enum values, indexed by matchable name.
+ * @param componentPropertyDefinitions
+ * @returns A map of {name: values[]}. Each value is an array to avoid
+ * collisions between properties / enum values
+ */
+export function buildMatchableNamesMap(
+  componentPropertyDefinitions?: FigmaRestApi.Component['componentPropertyDefinitions'],
+) {
+  const matchableValues: Record<string, MatchableName[]> = {}
+
+  function add(matchableStr: string, definition: MatchableName) {
+    matchableValues[matchableStr] = matchableValues[matchableStr] || []
+    matchableValues[matchableStr].push(definition)
+  }
+
+  Object.entries(componentPropertyDefinitions || {}).forEach(([propName, propDef]) => {
+    const matchableStr = stripNodeIdFromPropertyName(propName).replace(DELIMITERS_REGEX, '')
+
+    add(matchableStr, {
+      type: MatchableNameTypes.Property,
+      name: propName,
+    })
+
+    if (propDef.type === FigmaRestApi.ComponentPropertyType.Variant) {
+      propDef.variantOptions?.forEach((variantOption) => {
+        const variantMatchableStr = variantOption.replace(DELIMITERS_REGEX, '')
+
+        add(variantMatchableStr, {
+          type: MatchableNameTypes.VariantValue,
+          name: variantOption,
+          variantProperty: propName,
+        })
+      })
+    }
+  })
+
+  return matchableValues
 }
 
 export function generatePropMapping({
@@ -94,35 +154,88 @@ export function generatePropMapping({
 }): PropMapping {
   const propMapping: PropMapping = {}
 
-  // Remove whitespace for closer matches with code props
-  const matchableFigmaProperties = Object.keys(componentPropertyDefinitions || {}).reduce(
-    (acc, propName) => {
-      const withoutWhitespace = propName.replace(/\s/g, '')
-      acc[withoutWhitespace] = propName
-      return acc
-    },
-    {} as Record<string, string>,
-  )
+  const matchableNames = buildMatchableNamesMap(componentPropertyDefinitions)
 
-  const searchSpace = Object.keys(matchableFigmaProperties)
+  const searchSpace = Object.keys(matchableNames)
   const searcher = new Searcher(searchSpace)
 
-  Object.entries(signature).forEach(([propName, propSignature]) => {
+  function findBestMatch(
+    matchData: MatchData<string>[],
+    { score, where }: { score: number; where: (item: MatchableName) => boolean },
+  ) {
+    for (const nameMatches of matchData) {
+      if (nameMatches.score < score) {
+        return null
+      }
+      const items = matchableNames[nameMatches.item]
+      const match = items.find(where)
+      if (match) {
+        return match
+      }
+    }
+    return null
+  }
+
+  for (const [propName, propSignatureWithOptionalModifier] of Object.entries(signature)) {
+    const propSignature = propSignatureWithOptionalModifier.startsWith('?')
+      ? propSignatureWithOptionalModifier.substring(1)
+      : propSignatureWithOptionalModifier
+
     const results = searcher.search(propName, { returnMatchData: true })
 
-    if (results.length > 0) {
-      const { item, score } = results[0]
-      const figmaPropName = matchableFigmaProperties[item]
-      const figmaPropDef = componentPropertyDefinitions[figmaPropName]
+    if (results.length === 0) {
+      continue
+    }
 
-      if (score > PROP_MINIMUM_MATCH_THRESHOLD) {
-        const intrinsic = generateIntrinsic({ propSignature, figmaPropName, figmaPropDef })
-        if (intrinsic) {
-          propMapping[propName] = intrinsic
+    /**
+     * First, look for matching property names with compatible types
+     */
+    const matchingProperty = findBestMatch(results, {
+      score: PROP_MINIMUM_MATCH_THRESHOLD,
+      where: (item) => item.type === MatchableNameTypes.Property,
+    })
+    if (matchingProperty) {
+      const { name: figmaPropName } = matchingProperty
+
+      const intrinsic = generateIntrinsic({
+        propSignature,
+        figmaPropName,
+        figmaPropDef: componentPropertyDefinitions[figmaPropName],
+      })
+      if (intrinsic) {
+        propMapping[propName] = intrinsic
+      }
+    }
+
+    /**
+     * Then if no match AND a boolean prop, look for matching variant values, e.g:
+     *
+     * disabled: figma.enum('State', {
+     *   Disabled: true,
+     * })
+     */
+    if (!propMapping[propName] && propSignature === 'false | true') {
+      const matchingProperty = findBestMatch(results, {
+        score: PROP_MINIMUM_MATCH_THRESHOLD,
+        where: (item) => item.type === MatchableNameTypes.VariantValue,
+      })
+      if (matchingProperty) {
+        const { name: variantValue, variantProperty } = matchingProperty
+        if (!variantProperty) {
+          throw new Error('Expected variant property') // satisfying TS
+        }
+        propMapping[propName] = {
+          kind: IntrinsicKind.Enum,
+          args: {
+            figmaPropName: variantProperty,
+            valueMapping: {
+              [variantValue]: true,
+            },
+          },
         }
       }
     }
-  })
+  }
 
   return propMapping
 }
@@ -139,7 +252,10 @@ export function extractSignatureAndGeneratePropMapping({
   projectInfo: ReactProjectInfo
   componentPropertyDefinitions: FigmaRestApi.Component['componentPropertyDefinitions']
   cmd: BaseCommand
-}): PropMapping | undefined {
+}): {
+  propMapping: PropMapping | undefined
+  signature: ComponentTypeSignature | undefined
+} {
   let signature: ComponentTypeSignature
 
   try {
@@ -148,15 +264,24 @@ export function extractSignatureAndGeneratePropMapping({
       sourceFilePath: filepath,
       projectInfo,
     })
+    if (cmd.verbose && Object.keys(signature).length === 0) {
+      logger.warn(`No TS signature found for "${exportName}" in ${filepath}`)
+    }
   } catch (e) {
     if (cmd.verbose) {
-      logger.warn(`Could not extract signature for "${exportName}" in ${filepath}`)
+      logger.warn(`Could not extract TS signature for "${exportName}" in ${filepath}`)
     }
-    return undefined
+    return {
+      propMapping: undefined,
+      signature: undefined,
+    }
   }
 
-  return generatePropMapping({
-    componentPropertyDefinitions,
-    signature,
-  })
+  return {
+    propMapping: generatePropMapping({
+      componentPropertyDefinitions,
+      signature,
+    }),
+    signature: signature,
+  }
 }